requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
httpx
python-multipart
twilio
pydantic-settings
